Fix Node.js web deployment flow and TypeScript build errors
Summary of changes:
1.  **API Schema & Backend**:
    *   Updated `api/openapi.yaml` to include a `labels` property in the `Task` schema.
    *   Modified `src/frontend/api/task.php` and `src/frontend/api/tasks.php` to extract and return GitHub labels for tasks.
2.  **Web Frontend**:
    *   Enabled `output: 'export'` in `web/next.config.ts` to generate the static files required by the SSH deployment workflow.
    *   Regenerated TypeScript types using `openapi-typescript`.
    *   Fixed multiple TypeScript errors in `web/src/app/tasks/[id]/page.tsx`:
        *   Resolved missing `labels` property on the `Task` type.
        *   Fixed incorrect argument type for `performAction` (changed from string to object).
        *   Added null-check for `log.created_at`.
    *   Refactored dynamic routes (`/projects/[id]` and `/tasks/[id]`) to separate client-side logic into `*View` components and added `generateStaticParams` to the parent Server Components to support static export.
3.  **Deployment Workflow**:
    *   Added `env: FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true` to `.github/workflows/deploy.yml` to resolve Node.js 20 deprecation warnings.

Stuck points:
Next.js `output: export` requires `generateStaticParams` for all dynamic routes. I implemented these with empty arrays to allow the build to proceed, though in a real environment these might need to be populated with actual IDs or handled via a different deployment strategy if paths are truly dynamic. I also had to separate 'use client' components from the route entry points to allow exporting `generateStaticParams`.
