+++ .github/workflows/deploy.yml
@@ -19,6 +19,8 @@ jobs:
     name: Deploy to Webserver
     runs-on: ubuntu-latest
     if: github.actor == github.repository_owner
+    env:
+      FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
