+++ web/src/app/layout.tsx
@@ -18,6 +18,8 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
+export const dynamicParams = false;
+
 export default function RootLayout({
   children,
 }: Readonly<{
