+++ web/src/app/projects/[id]/ProjectDetailView.tsx
@@ -1,0 +1,298 @@
+'use client';
+
+import React, { use, useState, useMemo } from 'react';
+import Link from 'next/link';
+import { useProject } from '@/hooks/useProject';
+import { useTasks } from '@/hooks/useTasks';
+import { useTemplates } from '@/hooks/useTemplates';
+import StatusBadge from '@/components/StatusBadge';
+import TaskFilterBar from '@/components/TaskFilterBar';
+import Navbar from '@/components/Navbar';
+import { components } from '@/types/api';
+
+type TaskStatus = components['schemas']['Task']['status'];
+
+export default function ProjectDetailView({ id }: { id: string }) {
+  const projectId = parseInt(id);
+  const { data: project, isLoading: projectLoading, error: projectError, syncIssues, isSyncing, createFromTemplate, isCreatingFromTemplate, createFromRoadmap, isCreatingFromRoadmap } = useProject(projectId);
+  const { data: tasks, isLoading: tasksLoading } = useTasks(projectId);
+  const { data: templates } = useTemplates();
+
+  const [search, setSearch] = useState('');
+  const [statusFilter, setStatusFilter] = useState<TaskStatus | 'all'>('all');
+
+  const filteredTasks = useMemo(() => {
+    if (!tasks) return [];
+    return tasks.filter((t) => {
+      const matchesSearch = (t.title?.toLowerCase().includes(search.toLowerCase()) ?? false) ||
+                           (t.issue_number?.toString().includes(search) ?? false);
+      const matchesStatus = statusFilter === 'all' || t.status === statusFilter;
+      return matchesSearch && matchesStatus;
+    });
+  }, [tasks, search, statusFilter]);
+
+  if (projectLoading) {
+    return (
+      <div className="flex items-center justify-center min-h-screen bg-gray-50">
+        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
+      </div>
+    );
+  }
+
+  if (projectError || !project) {
+    return (
+      <div className="flex flex-col items-center justify-center min-h-screen p-4 bg-gray-50">
+        <div className="bg-red-50 border border-red-200 text-red-800 px-4 py-3 rounded-lg max-w-md w-full">
+          <p className="font-bold">Error loading project</p>
+          <p className="text-sm">Please try again later or check your connection.</p>
+        </div>
+        <Link href="/" className="mt-4 text-blue-600 hover:underline">
+          Back to Dashboard
+        </Link>
+      </div>
+    );
+  }
+
+  return (
+    <div className="min-h-screen bg-gray-50 pb-12">
+      <Navbar />
+
+      <main className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 pt-8">
+        <div className="flex flex-col md:flex-row md:items-center justify-between mb-8 gap-4">
+          <div>
+            <nav className="flex mb-2" aria-label="Breadcrumb">
+              <ol className="flex items-center space-x-2 text-sm text-gray-500">
+                <li><Link href="/" className="hover:text-gray-700">Dashboard</Link></li>
+                <li><svg className="w-4 h-4" fill="currentColor" viewBox="0 0 20 20"><path fillRule="evenodd" d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z" clipRule="evenodd"></path></svg></li>
+                <li className="font-medium text-gray-900 truncate max-w-[200px]">{project.github_repo}</li>
+              </ol>
+            </nav>
+            <h2 className="text-3xl font-bold text-gray-900">{project.github_repo}</h2>
+            <p className="text-gray-500 text-sm mt-1">
+              Linked as <span className="font-medium">{project.github_username}</span>
+            </p>
+          </div>
+          <div className="flex space-x-3">
+            <button
+              onClick={() => syncIssues()}
+              disabled={isSyncing}
+              className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50 transition-all"
+            >
+              <svg className={`w-4 h-4 mr-2 ${isSyncing ? 'animate-spin' : ''}`} fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
+              </svg>
+              {isSyncing ? 'Syncing...' : 'Sync Issues'}
+            </button>
+            <Link
+              href={`/projects/${id}/settings`}
+              className="inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md shadow-sm text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 transition-all"
+            >
+              Settings
+            </Link>
+          </div>
+        </div>
+
+        <div className="grid grid-cols-1 lg:grid-cols-4 gap-8">
+          <div className="lg:col-span-3">
+            <TaskFilterBar
+              search={search}
+              onSearchChange={setSearch}
+              statusFilter={statusFilter}
+              onStatusFilterChange={setStatusFilter}
+            />
+
+            <div className="bg-white border border-gray-200 rounded-xl shadow-sm overflow-hidden">
+              <table className="min-w-full divide-y divide-gray-200">
+                <thead className="bg-gray-50">
+                  <tr>
+                    <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Issue</th>
+                    <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Status</th>
+                    <th scope="col" className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">Actions</th>
+                  </tr>
+                </thead>
+                <tbody className="bg-white divide-y divide-gray-200">
+                  {tasksLoading ? (
+                    <tr>
+                      <td colSpan={3} className="px-6 py-12 text-center text-gray-500 italic">
+                        <div className="flex justify-center"><div className="animate-spin rounded-full h-6 w-6 border-b-2 border-blue-500"></div></div>
+                      </td>
+                    </tr>
+                  ) : filteredTasks.length === 0 ? (
+                    <tr>
+                      <td colSpan={3} className="px-6 py-12 text-center text-gray-500 italic">
+                        No tasks found matching your filters.
+                      </td>
+                    </tr>
+                  ) : (
+                    filteredTasks.map((task) => (
+                      <tr key={task.id} className="hover:bg-gray-50 transition-colors">
+                        <td className="px-6 py-4">
+                          <div className="flex flex-col">
+                            <Link href={`/tasks/${task.id}`} className="text-sm font-bold text-blue-600 hover:underline">
+                              #{task.issue_number} - {task.title}
+                            </Link>
+                            <span className="text-xs text-gray-500 mt-1 line-clamp-1">
+                              {task.body?.substring(0, 100)}
+                              {(task.body?.length ?? 0) > 100 ? '...' : ''}
+                            </span>
+                          </div>
+                        </td>
+                        <td className="px-6 py-4 whitespace-nowrap">
+                          <StatusBadge status={task.status} />
+                        </td>
+                        <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                          <Link href={`/tasks/${task.id}`} className="text-blue-600 hover:text-blue-900">
+                            Details
+                          </Link>
+                        </td>
+                      </tr>
+                    ))
+                  )}
+                </tbody>
+              </table>
+            </div>
+          </div>
+
+          <div className="lg:col-span-1 space-y-6">
+            <div className="bg-white border border-gray-200 rounded-xl shadow-sm p-6">
+              <h3 className="text-lg font-bold text-gray-900 mb-4">Roadmap</h3>
+              {project.roadmap_data && project.roadmap_data.length > 0 ? (
+                <ul className="space-y-4">
+                  {project.roadmap_data.map((file, index: number) => (
+                    <li key={index} className="flex flex-col space-y-1">
+                      <div className="flex items-center justify-between group">
+                        <a
+                          href={file.html_url}
+                          target="_blank"
+                          rel="noopener noreferrer"
+                          className="text-sm text-blue-600 hover:underline flex items-center"
+                        >
+                          <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
+                          </svg>
+                          <span className="truncate max-w-[120px]">{file.name}</span>
+                        </a>
+                        {file.next_task && (
+                          <button
+                            onClick={() => file.name && createFromRoadmap(file.name)}
+                            disabled={isCreatingFromRoadmap}
+                            className="p-1 text-green-600 hover:bg-green-50 rounded transition-colors"
+                            title="Implement next step"
+                          >
+                            <svg className="w-4 h-4" fill="currentColor" viewBox="0 0 20 20"><path d="M4.5 2.691l11 6.309-11 6.309V2.691z"/></svg>
+                          </button>
+                        )}
+                      </div>
+                      {file.next_task && (
+                        <div className="pl-6">
+                          <div className="text-[10px] text-gray-500 italic bg-gray-50 rounded px-2 py-1 border border-gray-100">
+                            🚧 {file.next_task}
+                          </div>
+                        </div>
+                      )}
+                    </li>
+                  ))}
+                </ul>
+              ) : (
+                <p className="text-sm text-gray-500 italic">No roadmap files found.</p>
+              )}
+              {project.roadmap_updated_at && (
+                <p className="text-[10px] text-gray-400 mt-4 border-t pt-2">
+                  Last updated: {new Date(project.roadmap_updated_at).toLocaleString()}
+                </p>
+              )}
+            </div>
+
+            <div className="bg-white border border-gray-200 rounded-xl shadow-sm p-6">
+              <h3 className="text-lg font-bold text-gray-900 mb-4">Create from Template</h3>
+              {templates && templates.length > 0 ? (
+                <TemplateForm
+                  templates={templates}
+                  onSubmit={(templateId, params) => createFromTemplate({ templateId, params })}
+                  isSubmitting={isCreatingFromTemplate}
+                />
+              ) : (
+                <p className="text-sm text-gray-500 italic">
+                  No templates found.{' '}
+                  <Link href="/templates" className="text-blue-600 hover:underline">
+                    Create one here.
+                  </Link>
+                </p>
+              )}
+            </div>
+          </div>
+        </div>
+      </main>
+    </div>
+  );
+}
+
+function TemplateForm({
+  templates,
+  onSubmit,
+  isSubmitting,
+}: {
+  templates: components['schemas']['Template'][];
+  onSubmit: (id: number, params: Record<string, string>) => void;
+  isSubmitting: boolean;
+}) {
+  const [selectedId, setSelectedId] = useState<number>(templates[0].id || 0);
+  const [params, setParams] = useState<Record<string, string>>({});
+
+  const selectedTemplate = templates.find((t) => t.id === selectedId);
+
+  const placeholders = useMemo(() => {
+    if (!selectedTemplate) return [];
+    const combined = (selectedTemplate.title_template || '') + ' ' + (selectedTemplate.body_template || '');
+    const matches = combined.match(/%\d+/g) || [];
+    return [...new Set(matches)].sort((a, b) => parseInt(a.slice(1)) - parseInt(b.slice(1)));
+  }, [selectedTemplate]);
+
+  return (
+    <form
+      onSubmit={(e) => {
+        e.preventDefault();
+        onSubmit(selectedId, params);
+      }}
+      className="space-y-4"
+    >
+      <div>
+        <label className="block text-xs font-medium text-gray-700 mb-1">Select Template</label>
+        <select
+          value={selectedId}
+          onChange={(e) => setSelectedId(parseInt(e.target.value))}
+          className="block w-full border border-gray-300 rounded-md shadow-sm p-2 text-sm"
+        >
+          {templates.map((t) => (
+            <option key={t.id} value={t.id}>
+              {t.name}
+            </option>
+          ))}
+        </select>
+      </div>
+
+      {placeholders.map((p) => (
+        <div key={p}>
+          <label className="block text-xs font-medium text-gray-700 mb-1">
+            {selectedTemplate?.parameter_config?.[p] || p}
+          </label>
+          <input
+            type="text"
+            required
+            value={params[p] || ''}
+            onChange={(e) => setParams({ ...params, [p]: e.target.value })}
+            className="block w-full border border-gray-300 rounded-md shadow-sm p-2 text-sm"
+          />
+        </div>
+      ))}
+
+      <button
+        type="submit"
+        disabled={isSubmitting}
+        className="w-full bg-green-600 text-white px-4 py-2 rounded-md text-sm font-medium hover:bg-green-700 disabled:opacity-50"
+      >
+        {isSubmitting ? 'Creating...' : 'Create Issue'}
+      </button>
+    </form>
+  );
+}
